@greg0ire 2022-11-08
Hi! `git rebase --exec vendor/bin/phpcbf` should fix some (but not all) cs issues :)