@NoiseByNorthwest 2022-11-26
It looks very close to the issue I've described here https://github.com/doctrine/orm/issues/8390
And I've shared a custom hydrator which fixes the useless "repeated rows hydration" with a simple caching logic.

Does my workaround make sense ? Should it be implemented into ORM ?